Deps: Extract kotlinx coroutines version to root build gradle
FYI: You will notice an additional 'experimentation' related
'kotlinxCoroutinesExperimentationVersion' version. This is because that
module is using an older version of Coroutines, the '1.4.2', that is,
comparing to the newer '1.6.4' version that the main modules are using.
As such, and in order to avoid behavior changes, two such version were
extracted to the root 'build.gradle' file with an accompanying TODO for
the future.