Deps: Remove androidx annotation dependency from crashlogging
FYI: After updating the 'dependency-analysis' plugin to '2.19.0', as
part of #272, this plugin started working better than before ('1.33.0')
and thus created this new advise that was previously ignored, since
indeed there is no 'import androidx.annotation.*' import on this
'crashlogging' module.

PS: The only such one 'import androidx.annotation.*' import exist on
the 'AutomatticTracks' module, because it (seems) that it is the only
module that is still using Java classes.

PR: https://github.com/Automattic/Automattic-Tracks-Android/pull/272

------------------------------------------------------------------------

This removal was suggested by the dependency analysis report, see below:

Advice for :crashlogging
Unused dependencies which should be removed:
  implementation 'androidx.annotation:annotation:1.7.0'